LayoutTests/inspector/styles/add-new-rule-invalid-selector-expected.txt

Tests that adding a new rule with invalid selector works as expected.

Text

Running: init

Running: keyframesRuleSelector
[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;
